Add book owner feature

+++ Controllers/BooksController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -5,6 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using library.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.EntityFrameworkCore;
 using library.Infrastructure;
 
 namespace library.Controllers
@@ -26,6 +28,13 @@ namespace library.Controllers
       return _db.Books.ToList();
     }
 
+    [Authorize]
+    [HttpGet("user")]
+    public ActionResult<IEnumerable<Book>> GetUserBooks()
+    {
+      return _db.Books.Where(book => book.OwnerId == GetUserId() && book.OwnerId != null).ToList();
+    }
+
     [HttpGet("{id}")]
     public ActionResult<Book> GetBookById(int id)
     {
@@ -42,6 +51,8 @@ namespace library.Controllers
     [HttpPost("")]
     public ActionResult<Book> PostBook(Book model)
     {
+      model.OwnerId = GetUserId();
+
       _db.Books.Add(model);
       _db.SaveChanges();
 
@@ -52,13 +63,15 @@ namespace library.Controllers
     [HttpPut("{id}")]
     public IActionResult PutBook(int id, Book model)
     {
-      if(model == null) {
-        return BadRequest();
+      var book = _db.Books.AsNoTracking().FirstOrDefault<Book>(book => book.Id == model.Id);
+
+      if(book == null) {
+        return NotFound();
       }
 
-      // if(!_books.Any(book => book.Id == id)) {
-      //   return NotFound();
-      // }
+      if(book.OwnerId != GetUserId()) {
+        return Forbid();
+      }
 
       _db.Books.Update(model);
       _db.SaveChanges();
@@ -76,10 +89,25 @@ namespace library.Controllers
         return NotFound();
       }
 
+      if(book.OwnerId != GetUserId()) {
+        return Forbid();
+      }
+
       _db.Books.Remove(book);
       _db.SaveChanges();
 
       return Ok(book);
+    }
+
+    private int? GetUserId() {
+      var UserIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+      if(!string.IsNullOrEmpty(UserIdClaim)) 
+      {
+        return Convert.ToInt32(UserIdClaim);
+      }
+
+      return null;
     }
   }
 }
